requirements.txt
PyQt6
pywinpty
pyte
qdarktheme
openai
python-dotenv
